+++ navigation-drawer/src/main/java/net/xpece/material/navigationdrawer/descriptors/NavigationItemDescriptor.java
@@ -8,7 +8,7 @@ public interface NavigationItemDescriptor {
    * Get unique item ID. This only makes sense for navigation items. Valid IDs are greater than 0.
    * @return
    */
-  long getId();
+  int getId();
 
   boolean isSticky();
 }
