+++ navigation-drawer/src/main/java/net/xpece/material/navigationdrawer/list/CompactNavigationListFragmentImpl.java
@@ -49,7 +49,7 @@ interface CompactNavigationListFragmentImpl {
 
     void setBackgroundAttr(@AttrRes int attr);
 
-    void setSelectedItem(long id);
+    void setSelectedItem(int id);
 
     void onInflate(Activity activity, AttributeSet attrs, Bundle savedInstanceState);
