+++ navigation-drawer/src/main/java/net/xpece/material/navigationdrawer/list/NavigationListFragment.java
@@ -86,7 +86,7 @@ public class NavigationListFragment extends Fragment implements NavigationListFr
     }
 
     @Override
-    public void setSelectedItem(long id) {
+    public void setSelectedItem(int id) {
         mDelegate.setSelectedItem(id);
     }
